feat: generate spec file for mainnet
