+++ xtask/Cargo.toml
@@ -5,5 +5,20 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tempfile = "3.10.1"
 xflags = "0.3.2"
+byteorder = "1.5.0"
+bytes = "1.9.0"
+hex.workspace = true
+hex-literal = "0.4.1"
+serde = { workspace = true, features = ["std"] }
+serde_json = { workspace = true, features = ["std"] }
+tempfile = "3.14.0"
+tokio = { workspace = true, features = ["rt-multi-thread"] }
+
+sc-client-api.workspace = true
+parity-scale-codec = { workspace = true, features = ["std"] }
+sc-service.workspace = true
+sp-core = { workspace = true, features = ["std"] }
+sp-crypto-hashing = { workspace = true, features = ["std"] }
+sp-runtime = { workspace = true, features = ["std"] }
+frame-remote-externalities.workspace = true
